src/requirements.txt
asgiref==3.2.3
cffi==1.13.2
cryptography==2.8
Django==3.0.2
django-extra-fields==2.0.2
django-filter==2.2.0
django-filters==0.2.1
django-rest-knox==4.1.0
djangorestframework==3.11.0
Pillow==7.0.0
pkg-resources==0.0.0
pycparser==2.19
PyMySQL==0.9.3
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
asgiref==3.2.3
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
Django==3.0.2
django-braces==1.14.0
django-cors-headers==3.2.1
django-extra-fields==2.0.2
django-filter==2.2.0
django-filters==0.2.1
django-oauth-toolkit==1.2.0
django-otp==0.7.5
django-rest-framework-social-oauth2==1.1.0
django-rest-knox==4.1.0
djangorestframework==3.11.0
idna==2.8
oauthlib==3.1.0
Pillow==7.0.0
pkg-resources==0.0.0
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.7
